fix of warnings and memory leak regarding interaction of response in lua pipeline

+++ ChillHttp/src/include/chill_config.h
@@ -26,7 +26,7 @@ typedef struct {
 	// ##### LOGGING #####
 
 	// ##### THREADS #####
-	size_t maxConcurrentThreads;
+	int maxConcurrentThreads;
 
 	// ##### PIPELINE #####
